collections-research-agent/code/requirements.txt
# --- Common / Mandatory ---
python-dotenv==1.0.1

# --- Agent Framework: a2a (StarletteApp) ---
a2a==0.10.6
click==8.1.7
uvicorn==0.29.0

# --- Agent LLM / MCP tools ---
langchain-mcp-adapters==0.1.17
langchain-openai==0.1.4
